+ add 64-bit version of factor()
+ macro protected code to check for divisibility problems in tdiv_large.c
+ bigprimes seems to be fully working now
(still need better 128-bit prp function than fermat, MR 128 isn't working, bpsw in-progress)

+++ ysieve/soe.h
@@ -193,7 +193,7 @@ typedef struct {
     int tstartup;
 
 	// start and stop for computing roots
-	uint32_t startid, stopid;
+	uint64_t startid, stopid;
 
 	// stuff for computing PRPs
 	mpz_t offset, lowlimit, highlimit, tmpz;
